requirements.txt
pycairo==1.23.0
PyGObject==3.42.2
rtoml==0.9.0
paho-mqtt==2.0.0
